refactor: project structure
Removed unused test_suite and old implementations. Introduced pytest for testing, replacing unittest, and added necessary dev dependencies. Migrated tripadvisor module setup to use __init__.py for improved clarity and maintainability.
